requirements.txt
click==8.1.7
cloudpickle==3.0.0
cmdstanpy==1.2.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
db-dtypes==1.2.0
debugpy==1.8.1
decorator==5.1.1
dm-tree==0.1.8
etuples==0.3.9
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
fastprogress==1.0.3
filelock==3.13.3
flatbuffers==24.3.25
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
gitpython==3.1.43
grpcio==1.62.1
grpcio-status==1.62.1
h5netcdf==1.3.0
h5py==3.10.0
holidays==0.46
huggingface_hub==0.22.2
idna==3.6
immutabledict==4.2.0
importlib-resources==6.4.0
ipykernel==6.29.3
ipython==8.22.2
ipython-autotime==0.3.2
ipython-genutils==0.2.0
jedi==0.19.1
jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.2.0
kiwisolver==1.4.4
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.7
msgpack-numpy==0.4.8
multipledispatch==1.0.0
munkres==1.1.4
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.3
numexpr==2.8.7
numpy==1.24.3
numpyro==0.13.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.0
plotly==5.20.0
prompt-toolkit==3.0.42
prophet==1.1.5
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure_eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydeck==0.9.1
pygments==2.17.2
pyparsing==3.0.9
pysocks==1.7.1
pyyaml==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.12.0
seaborn==0.12.2
setuptools==69.2.0
six==1.16.0
smmap==5.0.1
stack_data==0.6.2
stanio==0.5.0
statsmodels==0.14.0
streamlit==1.37.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing-extensions==4.11.0
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
xarray==2024.3.0
xarray-einstats==0.7.0
zipp==3.17.0
pymc_marketing @ git+https://github.com/nazarmaidanenko/pymc-marketing.git